==> 05_fuzzywuzzy_string_matching.ipynb <==
# Capstone: Sephora. Predicting prices based on Ingredients

## Problem description

It is an assumption customers make that their skin care product price is dependent on the ingredients in this product. The goal of my projects is to see if I can predict prices of the products based on the ingredients. To accomplish this goal, I first had to gather my data. I used Sephora.com data for this.

### Project Structure:
- Notebook 0. Selenium URL Collection
- Notebook 1. Saving data from URL to an HTML file
- Notebook 2. Collecting Product Data
- Notebook 3. Data Cleaning 
- Notebook 4. EDA
- Notebook 5. Fuzzy String Matching
- Notebook 6. Regression Modeling
- Notebook 7. Classification Modeling

import pandas as pd
from fuzzywuzzy import fuzz
from itertools import groupby
from operator import itemgetter

products = pd.read_csv('./data/products_clean_eda.csv')

products.head(2)

all_ingredients = []
#this code goes through all of the rows of ingredients and combines them in to one list
for i in products.index:
    list_ingredients = products.ingredients[i].split(', ') #splits a row ingredients into a list 
    for j in list_ingredients:     
        all_ingredients.append(j.strip().lower()) 

ingredients = pd.DataFrame([(x,all_ingredients.count(x)) for x in set(all_ingredients)],
                           columns = ['ingredient', 'count']).sort_values(by='ingredient')

ingredients = ingredients[ingredients.ingredient != '']
ingredients = ingredients[ingredients.ingredient != '1']

#https://stackoverflow.com/questions/17366788/python-split-list-based-on-first-character-of-word
letters = []
for letter, words in groupby(sorted(ingredients.ingredient), key=itemgetter(0)):
    for word in words:
        letters.append(letter)
letters

#first letter of the ingredient name
ingredients['letters'] = letters
ingredients.head()

#empty list that the rations will go in to
fuzzy = []
#looping through the letters of the alphabet
for i in sorted(set(ingredients['letters'])):
    #when ingredients starts with that letter
    df = ingredients[ingredients['letters'] == i]
    #calculating ratios 
    for k in df.ingredient:
        for j in df.ingredient:
            fuzzy.append((k, j, fuzz.ratio(k, j), df[df.ingredient == k]['count'].values[0],
                          df[df.ingredient == j]['count'].values[0]))
#creating a dataframe of ratios
fuzzy_ratio = pd.DataFrame(fuzzy, columns = ['ingredient1', 'ingredients2', 'ratio', 'word1 count', 'word2 count'])
fuzzy_ratio

#only looking at top 80 ratio
fuzzy_ratio = fuzzy_ratio[fuzzy_ratio.ratio != 100]
top80=fuzzy_ratio[fuzzy_ratio.ratio > 80]
top80.head()

#checking for duplicates if the order of column is just reversed

#https://stackoverflow.com/questions/24676705/pandas-drop-duplicates-if-reverse-is-present-between-two-columns
top80['check_string'] = top80.apply(lambda row: ''.join(sorted([row['ingredient1'], row['ingredients2']])), axis=1)

#https://stackoverflow.com/questions/12497402/python-pandas-remove-duplicates-by-columns-a-keeping-the-row-with-the-highest
top80_nodupes = top80.groupby('check_string', group_keys=False).apply(lambda x: x.loc[x['word1 count'].idxmax()])
top80_nodupes

#pivot table represents all the ingredients to keep vs the ones to change
pivot = top80_nodupes.pivot(index = 'ingredient1', columns = 'ingredients2', values = 'word1 count')
pivot

#based on the pivot table create a dataframe with columns to replace and value to replace with
ingredient_replace = []
for i in pivot.columns:
    #choosing the maximum number of counts in pivot column to replace with
    ingredient_replace.append((i, pivot[pivot[i] == pivot.loc[:, i].max()].index.values[0]))
ingredient_replace_df = pd.DataFrame(ingredient_replace, columns = ['to_replace', 'replace_with'])
ingredient_replace_df

#need to replace now in the full data set
for i in products.index:
    split_list = products.loc[i, 'ingredients'].split(', ')
    for index, item in enumerate(split_list):
        if item in ingredient_replace_df['to_replace'].values:
            split_list[index] = ingredient_replace_df[ingredient_replace_df['to_replace']==item]['replace_with'].values[0]
    products.loc[i, 'ingredients'] = ', '.join(split_list)


products.to_csv('./data/products_fuzzywuzzy.csv', index = False)
